Update dietpi-install.sh
Adds an image to the VM description and aligns the links horizontally. 

Just a little html not over complicating it, the result is below.

What do you think? This is just an idea, It cleans up the VM summary page and allows the dietpi logo to be shown on the VM summary page.

If you have a logo that would work better and would like to use that feel free to change the link 👍