Set OS type to Linux 2.6+
Sets the OS type to Linux kernel 2.6+

(no idea how i missed this but better late than never 🤣)

Thinking of also adding `qemu-guest-agent` install but that would require booting the VM immediately. Would this work because DietPi has a setup process on first boot? Curious because maybe it can be added for the user.